extracted verify verification code logic to cognito.ts
